restructure directory. rename files and folder
